Added c302 for reference
Going to start optimizing and improving general code structure. I think we have the brain at a good point, though I still don't know why it reacts weirdly to the food, and we may be able to start toying with evolution in order to evolve a brain that will be attracted to the food.